reorganize and add in redux state
