+++ src/calculate/actions/getCalculatedScore.js
