+++ src/index.js
@@ -1,0 +1,6 @@
+import {createStore} from 'redux';
+import rootReducer from './rootReducer';
+
+export default () => {
+  return createStore(rootReducer);
+};
