--- src/reducers/AppReducer.js
@@ -1,17 +1,0 @@
-import {combineReducers} from 'redux';
-
-const INITIAL_STATE = {
-  current: [],
-  possible: ['Alice', 'Bob', 'Sammy'],
-};
-
-const friendsReducer = (state = INITIAL_STATE, action) => {
-  switch (action.type) {
-    default:
-      return state;
-  }
-};
-
-export default combineReducers({
-  friends: friendsReducer,
-});
