+++ src/rootReducer.js
@@ -1,0 +1,6 @@
+import {combineReducers} from 'redux';
+import CalculateReducer from './calculate/reducers/CalculateReducer';
+
+export default combineReducers({
+  calculate: CalculateReducer,
+});
